chore: update to maturin's recommended project layout for rust/python projects
The previous layout leads to an import error when installing with `maturin build` and `pip install .`.

This error was common enough that `maturin` changed the recommended project layout to what this commit does.

A prior PR attempted to solve this by altering `lib.name` in Cargo.toml, but that did not work for me.

- [Prior PR](https://github.com/apache/datafusion-python/pull/694)
- [maturin ImportError issue](https://github.com/PyO3/maturin/issues/490)
- [maturin changes recommended project structure](https://github.com/PyO3/maturin/pull/855)